moved ReadRowsIterator into _read_rows
